+++ covid_moonshot/cli.py
@@ -2,7 +2,7 @@ import functools
 import json
 import logging
 import fire
-from . import analyze_runs
+from .core import analyze_runs
 
 
 @functools.wraps(analyze_runs)
